Remove achievement support, this api was removed in Bukkit 1.15
