+++ src/main/java/com/ensifera/animosity/craftirc/CraftIRCListener.java
@@ -7,6 +7,7 @@ import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.event.player.AsyncPlayerChatEvent;
+import org.bukkit.event.player.PlayerAdvancementDoneEvent;
 import org.bukkit.event.player.PlayerCommandPreprocessEvent;
 import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.event.player.PlayerKickEvent;
@@ -168,6 +169,31 @@ final class CraftIRCListener implements Listener {
         msg.setField("realSender", event.getEntity().getName());
         msg.setField("prefix", this.plugin.getPrefix(event.getEntity()));
         msg.setField("suffix", this.plugin.getSuffix(event.getEntity()));
+        msg.post();
+    }
+
+    @EventHandler(priority = EventPriority.MONITOR)
+    public void onPlayerAdvancement(PlayerAdvancementDoneEvent event) {
+        if (this.plugin.isHeld(CraftIRC.HoldType.ADVANCEMENTS)) {
+            return;
+        }
+        if (event.getAdvancement() == null) {
+            return;
+        }
+        final RelayedMessage msg = this.plugin.newMsg(this.plugin.getEndPoint(this.plugin.cMinecraftTag()), null, "achievement");
+        if (msg == null) {
+            return;
+        }
+        if (event.getAdvancement().getKey().getKey().startsWith("recipes/")) {
+            return;
+        }
+        msg.setField("sender", event.getPlayer().getDisplayName());
+        msg.setField("message", event.getAdvancement().getKey().getKey());
+        msg.setField("world", event.getPlayer().getWorld().getName());
+        msg.setField("realSender", event.getPlayer().getName());
+        msg.setField("prefix", this.plugin.getPrefix(event.getPlayer()));
+        msg.setField("suffix", this.plugin.getSuffix(event.getPlayer()));
+        msg.doNotColor("message");
         msg.post();
     }
 }
